requirements.txt
playwright>=1.40.0
twilio>=8.0.0
python-dotenv>=1.0.0
